i#3092 genapi: Split dr_os_utils+dr_tracedump from dr_tools
Multiple headers fed into dr_tools.h, which is no longer possible with
the simple copy-as-is export model.  We split the os-specific state,
aux libs, and memory query types from os_shared.h into a new file
os_api.h locally and dr_os_utils.h exported, and move the
corresponding routines from instrument_api.h into the file as well.

The tracedump types are similarly separated into fragment_api.h,
exported as dr_tracedump.h.

Furthermore, the rest of the dr_tools.h content is moved out of
instrument_api.h to core/lib/dr_tools.h, except process-query routines
which fit better in os_api.h.  This completes the refactoring of
dr_tools.h which is now completely removed from genapi.pl.

Issue: #3092
